Extract some of the apiDocs content into category index files
