+++ src/apiDefs/deprecated.test.ts
@@ -2,7 +2,7 @@ import 'jest';
 import * as moment from 'moment';
 
 // we just need a Markdown component for  our test IApiDescription
-import { FhirApiReleaseNotes } from '../content/apiDocs';
+import { FhirApiReleaseNotes } from '../content/apiDocs/health';
 import { IApiDescription } from './schema';
 
 jest.mock('./query');
