+++ src/content/apiDocs/benefits/index.ts
@@ -1,0 +1,11 @@
+import AppealsReleaseNotes from './appealsReleaseNotes.mdx';
+import BenefitsIntakeReleaseNotes from './benefitsIntakeReleaseNotes.mdx';
+import ClaimsReleaseNotes from './claimsReleaseNotes.mdx';
+import LoanGuarantyReleaseNotes from './loanGuarantyReleaseNotes.mdx';
+
+export {
+  AppealsReleaseNotes,
+  BenefitsIntakeReleaseNotes,
+  ClaimsReleaseNotes,
+  LoanGuarantyReleaseNotes,
+};
