+++ src/content/apiDocs/health/index.ts
@@ -1,0 +1,25 @@
+import HealthArgonautPostDeprecation from './argonautDeprecatedNotice.mdx';
+import HealthArgonautPreDeprecation from './argonautDeprecationNotice.mdx';
+import ArgonautReleaseNotes from './argonautReleaseNotes.mdx';
+import CommunityCareApiIntro from './communityCareApiIntro.mdx';
+import CommunityCareReleaseNotes from './communityCareReleaseNotes.mdx';
+import FhirApiReleaseNotes from './fhirAPIReleaseNotes.mdx';
+import FhirArgonautApiIntro from './fhirArgonautApiIntro.mdx';
+import FhirDSTU2ApiIntro from './fhirDSTU2ApiIntro.mdx';
+import FhirR4ApiIntro from './fhirR4ApiIntro.mdx';
+import UrgentCareApiIntro from './urgentCareApiIntro.mdx';
+import UrgentCareReleaseNotes from './urgentCareReleaseNotes.mdx';
+
+export {
+  ArgonautReleaseNotes,
+  CommunityCareApiIntro,
+  CommunityCareReleaseNotes,
+  FhirApiReleaseNotes,
+  FhirArgonautApiIntro,
+  FhirDSTU2ApiIntro,
+  FhirR4ApiIntro,
+  HealthArgonautPostDeprecation,
+  HealthArgonautPreDeprecation,
+  UrgentCareApiIntro,
+  UrgentCareReleaseNotes,
+};
