+++ src/content/apiDocs/verification/index.ts
@@ -1,0 +1,9 @@
+import AddressValidationReleaseNotes from './addressValidationReleaseNotes.mdx';
+import VeteranConfirmationReleaseNotes from './veteranConfirmationReleaseNotes.mdx';
+import VeteranVerificationReleaseNotes from './veteranVerificationReleaseNotes.mdx';
+
+export {
+  AddressValidationReleaseNotes,
+  VeteranConfirmationReleaseNotes,
+  VeteranVerificationReleaseNotes,
+};
